// FE/src/api/models/createOrganizationDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SIGREF.API
 * OpenAPI spec version: 1.0
 */
import type { IdentifierDto } from "./identifierDto";
import type { CodeableConceptDto } from "./codeableConceptDto";
import type { ExtendedContactDetailDto } from "./extendedContactDetailDto";
import type { ReferenceDto } from "./referenceDto";

export interface CreateOrganizationDto {
  /** @nullable */
  identifier?: IdentifierDto[] | null;
  /** @nullable */
  active?: boolean | null;
  /** @nullable */
  type?: CodeableConceptDto[] | null;
  /** @minLength 1 */
  name: string;
  /** @nullable */
  alias?: string[] | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  contact?: ExtendedContactDetailDto[] | null;
  partOf?: ReferenceDto;
  /** @nullable */
  endpoint?: ReferenceDto[] | null;
}